+++ cmd/packageloader/.gitignore
@@ -1,3 +1,4 @@
 cover.html
 cover.out
 packageloader
+encoder/test
