+++ cmd/packageloader/.vscode/configurationCache.log
@@ -1,0 +1,1 @@
+{"buildTargets":["build","test_encoder","test_loader"],"launchTargets":[],"customConfigurationProvider":{"workspaceBrowse":{"browsePath":[],"compilerArgs":[]},"fileIndex":[]}}
