+++ cmd/packageloader/.vscode/dryrun.log
@@ -1,0 +1,5 @@
+make --dry-run --always-make --keep-going --print-directory
+make: Entering directory '/home/nishyda/Documents/code/go/cx/cmd/packageloader'
+go build -o packageloader
+make: Leaving directory '/home/nishyda/Documents/code/go/cx/cmd/packageloader'
+
