+++ cmd/packageloader/database/database.go
@@ -1,4 +1,4 @@
-package server
+package database
 
 import (
 	"context"
