+++ cmd/packageloader/encoder/encoder.go
@@ -4,8 +4,8 @@ import (
 	"log"
 	"os"
 
+	"github.com/skycoin/cx/cmd/packageloader/database"
 	"github.com/skycoin/cx/cmd/packageloader/loader"
-	"github.com/skycoin/cx/cmd/packageloader/server"
 )
 
 func SavePackagesToDisk(packageName string, path string) {
@@ -14,15 +14,16 @@ func SavePackagesToDisk(packageName string, path string) {
 		log.Fatal(err)
 	}
 	var packageList loader.PackageList
-	packageList.UnmarshalBinary([]byte(server.Get(packageName).(string)))
+	packageList.UnmarshalBinary([]byte(database.Get(packageName).(string)))
 
 	for _, pack := range packageList.Packages {
 		var packageStruct loader.Package
-		packageStruct.UnmarshalBinary([]byte(server.Get(pack).(string)))
+		packageStruct.UnmarshalBinary([]byte(database.Get(pack).(string)))
 
 		if packageStruct.PackageName != "main" {
 			continue
 		}
+
 		path = path + "src/"
 		var filePath = path
 
@@ -31,13 +32,12 @@ func SavePackagesToDisk(packageName string, path string) {
 
 	for _, pack := range packageList.Packages {
 		var packageStruct loader.Package
-		packageStruct.UnmarshalBinary([]byte(server.Get(pack).(string)))
+		packageStruct.UnmarshalBinary([]byte(database.Get(pack).(string)))
 
 		if packageStruct.PackageName == "main" {
 			continue
 		}
 		var filePath = path + packageStruct.PackageName + "/"
-
 		SaveFilesToDisk(packageStruct, filePath)
 	}
 
@@ -50,7 +50,7 @@ func SaveFilesToDisk(packageStruct loader.Package, filePath string) {
 	}
 	for _, file := range packageStruct.Files {
 		var fileStruct loader.File
-		fileStruct.UnmarshalBinary([]byte(server.Get(file).(string)))
+		fileStruct.UnmarshalBinary([]byte(database.Get(file).(string)))
 
 		err = os.WriteFile(filePath+fileStruct.FileName, fileStruct.Content, 0755)
 		if err != nil {
