+++ cmd/packageloader/encoder/encoder_test.go
@@ -4,5 +4,4 @@ import "testing"
 
 func TestSavePackage(t *testing.T) {
 	SavePackagesToDisk("Test2", "../encoder/test/")
-
 }
