--- cmd/packageloader/loader/cover.out
@@ -1,82 +1,0 @@
-mode: set
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:38.51,39.28 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:44.2,44.14 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:39.28,40.22 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:40.22,42.4 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:47.13,48.39 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:53.2,56.86 3 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:65.2,65.16 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:69.2,69.37 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:72.2,73.33 2 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:48.39,50.3 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:50.8,52.3 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:56.86,57.17 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:60.3,60.19 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:63.3,63.13 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:57.17,59.4 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:60.19,62.4 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:65.16,67.3 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:69.37,71.3 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:73.33,75.3 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:78.60,79.31 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:82.2,83.50 2 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:86.2,90.16 5 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:93.2,93.33 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:99.2,99.24 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:110.2,110.23 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:123.2,126.31 4 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:79.31,81.3 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:83.50,85.3 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:90.16,92.3 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:93.33,94.55 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:97.3,97.40 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:94.55,95.12 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:99.24,101.17 2 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:104.3,105.17 2 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:108.3,108.39 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:101.17,103.4 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:105.17,107.4 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:110.23,114.17 4 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:117.3,117.19 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:121.3,121.39 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:114.17,116.4 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:117.19,120.4 2 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:126.31,127.36 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:130.3,130.45 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:127.36,128.12 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:135.100,137.16 2 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:140.2,141.16 2 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:144.2,144.37 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:157.2,157.40 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:137.16,139.3 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:141.16,143.3 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:144.37,146.17 2 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:149.3,150.17 2 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:153.3,153.36 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:146.17,148.4 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:150.17,152.4 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:153.36,155.4 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:161.59,163.16 2 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:166.2,169.34 4 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:175.2,176.28 2 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:163.16,165.3 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:169.34,171.115 2 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:171.115,173.4 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:180.75,182.16 2 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:185.2,188.21 4 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:197.2,197.21 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:182.16,184.3 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:188.21,189.33 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:193.3,193.115 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:189.33,192.4 2 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:193.115,194.9 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:201.61,202.32 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:202.32,208.17 3 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:211.3,214.32 4 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:208.17,210.4 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:219.58,224.16 4 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:227.2,230.12 3 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:224.16,226.3 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:234.72,239.16 4 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:242.2,244.12 3 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:239.16,241.3 1 0
