+++ cmd/packageloader/main.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+	"flag"
+	"fmt"
 	"log"
 	"os"
 
@@ -9,28 +11,37 @@ import (
 )
 
 func main() {
-	if len(os.Args) < 4 {
-		log.Fatal("Not enough arguments given. Usage: packageloader <option> <program name> <path>")
+	loadFlag := flag.Bool("load", false, "OPTION: Load a program to the database, with a given name to load it as and path to the program")
+	saveFlag := flag.Bool("save", false, "OPTION: Save a package to disk, with a given name to search on the database and a new directory path to save to")
+	helpFlag := flag.Bool("help", false, "OPTION: Display this help message")
+	nameFlag := flag.String("name", "", "The name of the program to load or save")
+	pathFlag := flag.String("path", "", "The path to the program to load or save")
+	flag.Parse()
+	if *helpFlag {
+		fmt.Println("Syntax: packageloader [OPTION] -path [PATH] -name [NAME]")
+		flag.Usage()
+		os.Exit(0)
 	}
-	if len(os.Args) > 4 {
-		log.Fatal("Too many arguments given. Usage: packageloader <option> <program name> <path>")
+	if flag.NFlag()+flag.NArg() > 3 || flag.NFlag()+flag.NArg() < 3 {
+		log.Fatal("Wrong number of arguments. Type -help for more information")
 	}
-	programName := os.Args[2]
 
-	var path string
-	if os.Args[3][0:2] == "./" {
-		path = os.Args[3][2:len(os.Args[3])]
-	} else {
-		path = os.Args[3]
-	}
+	programName := *nameFlag
+	path := *pathFlag
 	if path[len(path)-1:] != "/" {
-		path = path + "/"
+		path += "/"
+	}
+	if path[0:2] == "./" {
+		path = path[2:]
 	}
 
-	if os.Args[1] == "-l" || os.Args[1] == "-load" {
+	if *loadFlag {
 		loader.LoadPackages(programName, path)
+		return
 	}
-	if os.Args[1] == "-s" || os.Args[1] == "-save" {
+	if *saveFlag {
 		encoder.SavePackagesToDisk(programName, path)
+		return
 	}
+	log.Fatal("Wrong arguments provided. Type -help for more information")
 }
